feat(mypage): match record 일반/랭크 각각의 결과 나옴
